[#17] Separation of server & service structure.
By separating the actual server instance class and the service
implementation classes, it will be easier to add new services to the
server by simply attaching the service to the server.
