+++ common/src/main/proto/gc_service.proto
@@ -5,7 +5,7 @@ option java_package = "edu.kaist.algo.client";
 
 // The client streams the file contents to the server,
 // and requests for the result status.
-service GcService {
+service LogUpload {
   rpc InfoUpload (FileInfo) returns (FileUploadResult) {}
   rpc LogUpload (stream UploadRequest) returns (UploadResult) {}
 }
